[BugFix][Event] The new event logic adapts to triggerEvent.
description:
1. Added the enableEventHandleRefactor switch to control the new event handling logic.
2. Adapted the triggerEvent interface to trigger CustomEvent.
3. Fixed an issue where catch interception failed.
4. Adjusted the event triggering order.

issue: m-6577415519

SkipChecks: macro

+++ core/event/touch_event.h
@@ -26,9 +26,9 @@ class TouchEvent : public Event {
 
   TouchEvent(const std::string& event_name, float x = 0, float y = 0,
              float page_x = 0, float page_y = 0, float client_x = 0,
-             float client_y = 0, float time_stamp = 0);
+             float client_y = 0, int64_t time_stamp = 0);
   TouchEvent(const std::string& event_name, const lepus::Value& targets_touches,
-             float time_stamp = 0);
+             int64_t time_stamp = 0);
 
   float x() const { return x_; }
   float y() const { return y_; }
